Practice Calendar Updates
Able to pull events below the calendar but it actually goes below the footer, I think I have to make a section that goes beneath the calendar and append in events there.

The next bug is that it crams all event Icons into the card for the day, I need one for each day. Perhaps its because All the icons are built and then put into the ev list at the the end.