Calendar Number Separation
Mobile Calendar events are no longer inside of the calendar cell and are being displayed seperatly as cards at the bottom.

The next challenge is to get each card to be their own instead of lumped by day. Something tells me this has something to do with how they are being filtered by day.